+++ src/index.tsx
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom";
 import "./index.css";
 import App from "./containers/App/App";
 import { Provider } from "react-redux";
-import { store } from "./store";
+import store from "./store";
 
 ReactDOM.render(
   <Provider store={store}>
